[geometry] Fix area geometry tests
